pyproject.toml
[project]
name = "knou-rate-course"
version = "0.1.0"
description = "방통대 꿀과목 DB - Course rating service for KNOU students"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
]
